only warn about bad config version if the file actually exists

+++ Desktop/src/org/mitre/opensextant/desktop/ui/helpers/ConfigHelper.java
@@ -114,22 +114,22 @@ public class ConfigHelper {
                 dataHome.mkdirs();
 
             File settingsFile = new File(CONFIG_FILE);
+            boolean settingsFileExisted = true;
             if (!settingsFile.exists()) {
+                settingsFileExisted = false;
                 settingsFile.createNewFile();
             }
             config = new PropertiesConfiguration(CONFIG_FILE);
             int version = config.getInt("version", 0);
-            if (version != VERSION) {
+            if (settingsFileExisted && version != VERSION) {
                 File configBackup = new File(CONFIG_FILE+"_v"+version);
                 File jobsBackup = new File(JOBS_FILE+"_v"+version);
                 
-                if (settingsFile.exists()) {
-                    FileUtils.copyFile(new File(CONFIG_FILE), configBackup);
-                    if (new File(JOBS_FILE).exists()) FileUtils.copyFile(new File(JOBS_FILE), jobsBackup);
-                    JOptionPane.showMessageDialog(null, "Your OpenSextant configuration files are from an incompatable version and cannot be used.  They have been backed up to: " + configBackup.getAbsolutePath() + ".");
-                    FileUtils.deleteQuietly(new File(CONFIG_FILE));
-                    FileUtils.deleteQuietly(new File(JOBS_FILE));
-                }
+                FileUtils.copyFile(new File(CONFIG_FILE), configBackup);
+                if (new File(JOBS_FILE).exists()) FileUtils.copyFile(new File(JOBS_FILE), jobsBackup);
+                JOptionPane.showMessageDialog(null, "Your OpenSextant configuration files are from an incompatable version and cannot be used.  They have been backed up to: " + configBackup.getAbsolutePath() + ".");
+                FileUtils.deleteQuietly(new File(CONFIG_FILE));
+                FileUtils.deleteQuietly(new File(JOBS_FILE));
                 settingsFile.createNewFile();
                 config = new PropertiesConfiguration(CONFIG_FILE);
             }
